requirements.txt
Django==3.2
djangorestframework==3.12.4
Pillow==8.2.0
